==> src/notebooks/1_Carga_Y_PrimerAnalisis.ipynb <==
## Importamos librerías necesarias

import pandas as pd
import requests, json
import time
import sys

pd.set_option("display.max_columns", None)

## Primer dataset 
### CSV con los resultados de los partidos del circuito ATP desde el año 2000

### Carga de datos

sys.path.append("../")
dfResultados = pd.read_csv("../data/raw/atp_tennis.csv", sep=";")
dfResultados.head()

dfResultados.info()

dfResultados.describe().round(2)

dfResultados.shape

### Primer análisis de la información

dfResultados.isna().sum()/len(dfResultados) *100 

Como no hay ningún valor NaN, no es necesario aplicar ningún tratamiento

dfResultados.to_csv("../data/process/resultados.csv")

## Segundo dataset 
### API que contiene estadísticas de jugadores

### Carga de datos

url = "https://ultimate-tennis1.p.rapidapi.com/live_leaderboard/50"

headers = {
	"X-RapidAPI-Key": "0b00932e03msh9c0e6292431999cp102f2cjsnfaa91f7ec449",
	"X-RapidAPI-Host": "ultimate-tennis1.p.rapidapi.com"
}

response = requests.get(url, headers=headers)

todos = json.loads(response.text)

players = pd.DataFrame(todos["data"])

ids = players[["Name","id"]]

dfStats = pd.DataFrame()

for i in range(len(ids)):
	id = ids.iloc[i]["id"]
	name = ids.iloc[i]["Name"]
	# print(id, name)
	url = (f"https://ultimate-tennis1.p.rapidapi.com/player_stats/atp/{id}/2023/all")

	headers = {
		"X-RapidAPI-Key": "0b00932e03msh9c0e6292431999cp102f2cjsnfaa91f7ec449",
		"X-RapidAPI-Host": "ultimate-tennis1.p.rapidapi.com"
	}

	response = requests.get(url, headers=headers)

	playerStats = json.loads(response.text)

	data1 = pd.DataFrame(playerStats["player_data"]).head()
	# print(data1)

	name = pd.DataFrame([name], columns=["Name"])
	conc = pd.concat([name, data1], axis=1)
	dfStats = pd.concat([dfStats, conc], ignore_index=True)
	time.sleep(5)
dfStats.head()

### Primer análisis de la información

dfStats.info()

dfStats.describe()

dfStats.to_csv("../data/process/estadisticas.csv")

<hr>
<div>
    <a href="./2_Preparacion_Datos.ipynb">
        <button style="float: right;">2.Preparación de datos &#8594;</button>
    </a>
</div>
<hr>
