==> src/notebooks/2_Preparacion_Datos.ipynb <==
## Importamos librerías necesarias

import pandas as pd
import sys

pd.set_option("display.max_columns", None)
sys.path.append("../")

## Primer dataset 

dfResultados = pd.read_csv("../data/process/resultados.csv")
dfResultados.head()

Eliminamos las columnas que no vamos a utilizar y convertimos el campo "Date" a tipo Date para un mejor tratamiento

#Eliminamos las columnas que no vamos a utilizar
dfResultados.drop(["Odd_1", "Odd_2", "Pts_1", "Pts_2", "Unnamed: 0"], axis=1, inplace=True)

#Convertimos el campo "Date" a tipo Date para un mejor tratamiento
dfResultados['Date'] = pd.to_datetime(dfResultados['Date'], format='%d/%m/%Y')
dfResultados.head()

Nos quedamos con los resultados de este año para hacer un análisis más personalizados

df2023 = dfResultados[dfResultados["Date"].dt.year == 2023].reset_index()
df2023.head()

Filtramos por los principales tipos de torneos del circuito ATP (Grand Slam, Masters Cup y Masters 1000).

Añadimos también la columna Sets_Played (que contiene el número de sets jugados por partido), y Tie_Breaks (que contiene el número de Tie Breaks jugados por partido)

torneos = ['Grand Slam', "Masters Cup", 'Masters 1000']
df_torneos = df2023[df2023["Series"].isin(torneos)]

df_torneos["Sets_Played"] = df_torneos["Score"].str.count("-")
df_torneos["Tie_Breaks"] = df_torneos["Score"].str.count("7-6") + df_torneos["Score"].str.count("6-7")
df_torneos.set_index("index", inplace=True)
df_torneos.head()

### Descripción de columnas de dataframe

| Columnas | Descripción |
|----------|----------|
| Tournament | Nombre del torneo
|Date|Fecha del partido
|Series|Categoría del torneo
|Court|Si se juega en interior o exterior
|Surface|Tipo de superficie donde se juega el torneo
|Best of|Al mejor de que sets se juega
|Player_1|Jugador 1
|Player_2|Jugador 2
|Winner|Ganador del partido
|Rank_1|Ranking del jugador 1 al comienzo de ese torneo
|Rank_2|Ranking del jugador 2 al comienzo de ese torneo
|Score|Puntuación del partido
|Num_Sets_Played|Número de sets jugados
|Tie_Breaks|Número de Tie-Breaks del partido

df_torneos.to_csv(r"../data/final/resultados_final.csv")

## Segundo dataset 

dfStats = pd.read_csv("../data/process/estadisticas.csv")
dfStats.head()

Los nombres de las columnas, tal y como venían de la API, entre cada palabra venían con un caracter "extraño", que no era ni salto de línea, ni tabulación... Por ello, lo primero que he hecho ha sido renombrarlas para poder usarlas mejor

dfStats = dfStats.set_axis(["Unnamed: 0","Player Name", "First serve", "First Serve Points Won", "1st Serve Return Points Won", "2nd Serve Points Won","2nd Serve Return Points Won",
                            "Aces","Break Points Converted", "Break Points Faced", "Break Points Opportunities", "Break Points Saved", "Double Faults",
                            "Return Games Played", "Return Games Won %", "Return Points Won", "Service Games Played", "Service Games Won %", "Total Points Won",
                            "Total Service Points Won"], axis=1)

dfStats.drop(["Unnamed: 0"], axis=1, inplace=True)
dfStats.head()

Al ser la mayoría de las columnas de tipo "object", para poder sacar algo de información las que contenían un % las he convertido a "float"

dfStats['First serve'] = dfStats['First serve'].str.rstrip('%').astype('float') / 100.0
dfStats['First Serve Points Won'] = dfStats['First Serve Points Won'].str.rstrip('%').astype('float') / 100.0
dfStats['2nd Serve Points Won'] = dfStats['2nd Serve Points Won'].str.rstrip('%').astype('float') / 100.0
dfStats['Break Points Converted'] = dfStats['Break Points Converted'].str.rstrip('%').astype('float') / 100.0
dfStats['Break Points Saved'] = dfStats['Break Points Saved'].str.rstrip('%').astype('float') / 100.0
dfStats['Return Games Won %'] = dfStats['Return Games Won %'].str.rstrip('%').astype('float') / 100.0
dfStats['Return Points Won'] = dfStats['Return Points Won'].str.rstrip('%').astype('float') / 100.0
dfStats['Service Games Won %'] = dfStats['Service Games Won %'].str.rstrip('%').astype('float') / 100.0
dfStats['Total Service Points Won'] = dfStats['Total Service Points Won'].str.rstrip('%').astype('float') / 100.0
dfStats['Total Points Won'] = dfStats['Total Points Won'].str.rstrip('%').astype('float') / 100.0
dfStats.head()

Al igual que en paso anterior, he convertido el resto de las que contenían un número a tipo "int64".

**¿Por qué hacerlo en dos pasos y no en uno solo?** En primer lugar, porque son tipos distintos, y en segundo por no tener una celda gigante

dfStats['Aces'] = dfStats['Aces'].str.replace(",","")
dfStats['Aces'] = dfStats['Aces'].astype("int64")
dfStats['Break Points Faced'] = dfStats['Break Points Faced'].astype("int64")
dfStats['Double Faults'] = dfStats['Double Faults'].astype("int64")
dfStats['Break Points Opportunities'] = dfStats['Break Points Opportunities'].astype("int64")
dfStats['Service Games Played'] = dfStats['Service Games Played'].str.replace(",","")
dfStats['Service Games Played'] = dfStats['Service Games Played'].astype("int64")
dfStats['Return Games Played'] = dfStats['Return Games Played'].str.replace(",","")
dfStats['Return Games Played'] = dfStats['Return Games Played'].astype("int64")
dfStats.head()

Como a pesar de que venía mucha información, me faltaban algunos datos he añadido columnas que informan de: 
- Puntos de break ganados ("BreakPoints Won")
- Puntos de break salvados ("BreakPoints Saved")
- Juegos al servicio ganados ("Service Games Won")
- Juegos al resto ganados ("Return Games Won")

dfStats["BreakPoints Won"] = (dfStats["Break Points Faced"] * dfStats["Break Points Converted"]).round()
dfStats["BreakPoints Saved"] = (dfStats["Break Points Opportunities"] * dfStats["Break Points Saved"]).round()
dfStats["Return Games Won"] = (dfStats["Return Games Played"] * dfStats["Return Games Won %"]).round()
dfStats["Service Games Won"] = (dfStats["Service Games Played"] * dfStats["Service Games Won %"]).round()
dfStats.head()

### Descripción de columnas de dataframe

| Columnas | Descripción |
|----------|----------|
| First Serve | Porcentaje de primeros servicios metidos
|First Serve Points Won|Puntos ganados (%) con el primer servicio
|1st Serve Return Points Won|Puntos ganados (%) con el primer servicio (al resto)
|2nd Serve Points Won|Puntos ganados (%) con el segundo servicio
|2nd Serve Return Points Won|Puntos ganados (%) con el segundo servicio (al resto)
|Aces|Saques directos
|Break Points Converted|Puntos de Break (%) ganados (al resto)
|Break Points Faced|Puntos de Break jugados (al resto)
|Break Points Opportunities|Puntos de Break jugados (al servicio)
|Break Points Saved|Puntos de Break salvados (al servicio) (%)
|Double Faults|Número de dobles faltas
|Return Games Played|Juegos al resto jugados
|Return Games Won %|Juegos al resto ganados (%)
|Return Points Won|Puntos ganados al resto (%)
|Service Games Played|Juegos al servicio jugados
|Service Games Won %|Juegos al servicio ganados (%)
|Total Points Won|Puntos totales ganados (%)
|Service Points Won|Puntos ganados al servicio (%)
|BreakPoints Won|Número de puntos de break ganados (al resto)
|BreakPoints Saved|Número de puntos de break ganados (al servicio)
|Return Games Won|Número de juegos al resto ganados
|Service Games Won|Número de juegos al servicio ganados

dfStats.to_csv(r"../data/final/estadisticas_final.csv")

<hr>
<div>
    <a href="./1_Carga_Y_PrimerAnalisis.ipynb">
        <button style="float: left;">&#8592; 1.Carga y primer análisis</button>
    </a>
    <a href="./3_Análisis.ipynb">
        <button style="float: right;">3.Análisis &#8594;</button>
    </a>
</div>
<hr>
