==> src/notebooks/4_Hipotesis.ipynb <==
## Importamos librerías necesarias

import pandas as pd

import scipy.stats as stats
import sys

pd.set_option("display.max_columns", None)
sys.path.append("../")

## Pre-cargamos datasets

df_torneos = pd.read_csv("../data/final/resultados_final.csv")
df_torneos['Date'] = pd.to_datetime(df_torneos['Date'])
df_torneos.head()

dfStats = pd.read_csv("../data/final/estadisticas_final.csv")
dfStats.drop("Unnamed: 0", axis=1, inplace=True)
dfStats.head()

## Hipótesis

### Hipótesis 1
- H0 -> No hay evidencia significativa de la correlación entre número de aces y el número de servicios ganados
- H1 -> Existe una evidencia significativa de la correlación entre número de aces y el número de servicios ganados

df_aces = dfStats["Aces"]
df_DobleFalta = dfStats["Service Games Won"]

correlacion , pv = stats.pearsonr(df_aces, df_DobleFalta)

if pv < 0.05:
    print('Rechazamos la hipótesis nula "No hay evidencia significativa de la correlación entre número de aces y el número de servicios ganados"')
else:
    print('No rechazamos la hipótesis nula')

### Hipótesis 2
- H0 -> No hay evidencia significativa de la relación entre número de dobles faltas y el número de Break Points salvados
- H1 -> Existe una evidencia significativa de la relación entre número de dobles faltas y el número de Break Points salvados

doblesFaltas = dfStats["Double Faults"]

breakPointSaved = dfStats["BreakPoints Saved"]

ttest = stats.ttest_ind(doblesFaltas, breakPointSaved)


if ttest[1] < 0.05:
    print('Rechazamos la hipótesis nula "No hay evidencia significativa de la relación entre número de dobles faltas y el número de Break Points salvados"')
else:
    print('No rechazamos la hipótesis nula')

### Hipótesis 3
- H0 -> La distribución de las columnas numéricas sigue una distribución normal
- H1 -> La distribución de las columnas numéricas NO sigue una distribución normal.

#4.
variables_num = dfStats.select_dtypes(exclude='object')

shapiro = {var: stats.shapiro(variables_num[var]) for var in variables_num.columns}

for var in variables_num.columns:
    if shapiro[var][1] < 0.05:
        print(f'Rechazamos la hipótesis nula "No hay evidencia significativa en la normalidad de la variable {var}"')
    else:
        print(f'No rechazamos la hipótesis nula "No hay evidencia significativa en la normalidad de la variable {var}"')

### Hipótesis 4
- H0 -> No hay evidencia significativa de la correlación entre número de TB jugados en hierba y el número de TB jugados en arcilla
- H1 -> Existe una evidencia significativa de la correlación entre número de TB jugados en hierba y el número de TB jugados en arcilla

partidosHierba = df_torneos[df_torneos["Surface"] == "Grass"]["Tie_Breaks"]

partidosClay = df_torneos[df_torneos["Surface"] == "Clay"]["Tie_Breaks"]

ttest = stats.ttest_ind(partidosHierba, partidosClay)

if ttest[1] < 0.05:
    print('Rechazamos la hipótesis nula "No hay evidencia significativa de la correlación entre número de TB jugados en hierba \ny el número de TB jugados en arcilla"')
else:
    print('No rechazamos la hipótesis nula')

### Hipótesis 5
- H0 -> No hay evidencia de diferencias significativas en el número de sets jugados en los Grand Slam, Masters 1000 y Copa de Maestros
- H1 -> Existe una evidencia de diferencias significativas en el número de sets jugados en los Grand Slam, Masters 1000 y Copa de Maestros

grupos_torneos = df_torneos.groupby('Series')['Sets_Played']
grandSlam = grupos_torneos.get_group('Grand Slam')
masters1000 = grupos_torneos.get_group('Masters 1000')
mastersCup = grupos_torneos.get_group('Masters Cup')

anova = stats.f_oneway(grandSlam, masters1000, mastersCup)
anova

if anova[1] < 0.05:
    print('Rechazamos la hipótesis nula "No hay evidencia de diferencias significativas en el número de sets jugados \nen los Grand Slam, Masters 1000 y Copa de Maestros"')
else:
    print('No rechazamos la hipótesis nula')

### Hipótesis 6
- H0 -> No hay evidencia significativa de la correlación entre número de sets jugados y el número de Tie Breaks jugados
- H1 -> Existe una evidencia significativa de la correlación entre número de sets jugados y el número de Tie Breaks jugados

sets = df_torneos["Sets_Played"]
tieBreaks = df_torneos["Tie_Breaks"]

correlacion , pv = stats.pearsonr(sets, tieBreaks)

if pv < 0.05:
    print('Rechazamos la hipótesis nula "No hay evidencia significativa de la correlación entre número de sets jugados y el número de Tie Breaks jugados"')
else:
    print('No rechazamos la hipótesis nula')

<hr>
<div>
    <a href="./3_Análisis.ipynb">
        <button style="float: left;">&#8592; 3. Análisis</button>
    </a>
</div>
<hr>
